fix: svg type definition for jsx files
